Add support for ANY route.
This allows a route to match any HTTP method so the downstream code handling the route can decide what to do.  While the user can easily match multiple methods on a single route, this allows users to use FastRoute simply as a fast router, decoupled from what methods it may or may not care about it.