Remove extension point on result
That class is also a data transfer object and has no behaviour for this
library.

Users would have to implement their own dispatchers in order to
instantiate a different implementation of the result and append
data/behaviour to it.

It's much easier to encapsulate that in the "handler", which can be
anything, and allow for the design of this library to be extremely
simple and focus on routing.

This also removes the BC-layer, in favour of adding a FC-layer in the
1.x branch.