Introduce caching API
Allowing users to vary the implementation of route caching according to
their needs.

This makes the file system based cache a bit more robust by using
exclusive locking and preventing reads of partial files.

We're also adding an APCu-based implementation, so that we can compare
FastRoute to HackRouting.

More info:

- https://github.com/azjezz/hack-routing
- https://github.com/azjezz/benchmark-php-routing