Introduce result objects
Using arrays for the dispatching result hinders type safety and forces
us to dynamically verify what has happened.

This introduces result objects with read-only public variables that are
backwards compatible with the array shapes we were using, providing more
flexibility and a type-based way to know if we have or not a match.

It's really important to mention that we aren't calling the constructor
with promoted read-only variables because of the average overhead of 60% on
dispatching operations. Introducing objects with public properties also
has an impact but it's significantly smaller (average overhead of 12%).

I do believe that having an object there helps - as previously requested
by several people.

Signed-off-by: Luís Cobucci <lcobucci@gmail.com>